pyproject.toml
[tool.poetry]
name = "rottn"
version = "0.1.0"
description = "Wander the Realm Of The Thousand Nightmares. Meet your inevitable doom. Hope is futile."
authors = ["pxlbrain"]

[tool.poetry.dependencies]
python = "^3.6"
pygase = "^0.3.0"
cocos2d = "^0.6.5"
Keras = "^2.2"
euclid3 = "^0.1.0"
appdirs = "^1.4"
numpy = "^1.16"
tensorflow = "^1.13"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pytest = "^4.4"
pydoc-markdown = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
